refactor(project): Remove babel dependencies.
Remove babel dependencies.
Use @std/esm to support "import" and "export"

BREAKING CHANGE: File structure has been changed.
