+++ index.mjs
@@ -1,0 +1,3 @@
+import PasswordGenerator from './lib/passwordgenerator';
+
+export default PasswordGenerator;
