+++ lib/passwordgenerator.mjs
@@ -4,7 +4,7 @@ import { diceware8k } from './diceware8k';
 const privatePrng = new WeakMap();
 const privateSymbolTable = new WeakMap();
 
-class PasswordGenerator {
+export default class PasswordGenerator {
   /**
    * The Constructor
    * @param {array} symbolTable - list of distinct symbols used to build the password
@@ -112,5 +112,3 @@ class PasswordGenerator {
     return x;
   }
 }
-
-export { PasswordGenerator as default };
